Add catalog facet search test·
